Refactor existing Views
