+++ frontend/src/main/kotlin/at/downdrown/diary/frontend/view/View.kt
@@ -1,9 +1,15 @@
 package at.downdrown.diary.frontend.view
 
-enum class View(val location: String) {
+import at.downdrown.diary.frontend.view.registration.RegistrationView
+import com.vaadin.flow.component.Component
 
-    Login("/login"),
-    Registration("/register"),
-    Main("/")
+enum class View(
+    val navigationTarget: Class<out Component?>,
+    val i18nKey: String
+) {
+
+    Login(LoginView::class.java, "login.pagetitle"),
+    Registration(RegistrationView::class.java, "registration.pagetitle"),
+    Today(TodayView::class.java, "nav.today")
 
 }
